Reorganize folders
`nlp.corenlp` instead of `corenlp.client`
